Use returns_nonnull as the example
returns_nonnull is newer than noreturn and will make it easier to show
pitfalls. noreturn is also a poor choice because it is included in the
C11 standard. A reader could be distracted by the question of why we
would use the GNU variant when it is officially standardized.

The author of this commit has waived all copyright and
related or neighboring rights to the extent permitted by
law as described by the CC0 1.0 Universal Public Domain
Dedication. You should have received a copy of the full
dedication along with this commit, typically as a file
named <CC0-1.0.txt>. If not, it may be available at
<https://creativecommons.org/publicdomain/zero/1.0/>.