Reference a slightly older version of Clang
This will help us show the #ifdef __GNUC__ pitfall. Clang 3.5.0 sneakily
supports returns_nonnull and won't barf.

The author of this commit has waived all copyright and
related or neighboring rights to the extent permitted by
law as described by the CC0 1.0 Universal Public Domain
Dedication. You should have received a copy of the full
dedication along with this commit, typically as a file
named <CC0-1.0.txt>. If not, it may be available at
<https://creativecommons.org/publicdomain/zero/1.0/>.